refactor: application client 저장 로직 클래스 분리
