+++ src/main/java/com/map/gaja/client/application/ClientSavingService.java
@@ -1,0 +1,62 @@
+package com.map.gaja.client.application;
+
+import com.map.gaja.client.domain.model.Client;
+import com.map.gaja.client.domain.model.ClientImage;
+import com.map.gaja.client.infrastructure.repository.ClientQueryRepository;
+import com.map.gaja.client.infrastructure.repository.ClientRepository;
+import com.map.gaja.client.presentation.dto.request.NewClientRequest;
+import com.map.gaja.client.presentation.dto.response.ClientOverviewResponse;
+import com.map.gaja.global.authentication.AuthenticationRepository;
+import com.map.gaja.group.application.util.GroupServiceHelper;
+import com.map.gaja.group.domain.model.Group;
+import com.map.gaja.group.domain.service.IncreasingClientService;
+import com.map.gaja.group.infrastructure.GroupRepository;
+import com.map.gaja.user.domain.model.User;
+import com.map.gaja.user.infrastructure.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import static com.map.gaja.client.application.ClientConvertor.*;
+
+@Service
+@RequiredArgsConstructor
+public class ClientSavingService {
+    private final ClientRepository clientRepository;
+    private final GroupRepository groupRepository;
+    private final IncreasingClientService increasingClientService;
+    private final AuthenticationRepository securityUserGetter;
+    private final UserRepository userRepository;
+
+    /**
+     * 이미지 없는 고객 등록
+     * @param clientRequest 고객 등록 요청 정보
+     * @return 만들어진 고객 ID
+     */
+    public ClientOverviewResponse saveClient(NewClientRequest clientRequest, String loginEmail) {
+        User user = userRepository.findByEmail(loginEmail);
+        Group group = GroupServiceHelper.findGroupByIdForUpdating(groupRepository, clientRequest.getGroupId());
+        Client client = dtoToEntity(clientRequest, group, user);
+        clientRepository.save(client);
+        increasingClientService.increaseByOne(group, securityUserGetter.getAuthority().get(0));
+        return entityToOverviewDto(client);
+    }
+
+    /**
+     * 이미지와 함께 고객 등록
+     *
+     * @param clientRequest 고객 등록 요청 정보
+     * @param loginEmail 요청을 보내는 사용자 이메일
+     * @return 만들어진 고객
+     */
+    public ClientOverviewResponse saveClientWithImage(NewClientRequest clientRequest, String loginEmail) {
+        User user = userRepository.findByEmail(loginEmail);
+        Group group = GroupServiceHelper.findGroupByIdForUpdating(groupRepository, clientRequest.getGroupId());
+        ClientImage clientImage = ClientImage.create(loginEmail, clientRequest.getClientImage());
+        Client client = dtoToEntity(clientRequest, group, user, clientImage);
+        client.updateImage(clientImage);
+
+        clientRepository.save(client);
+        increasingClientService.increaseByOne(group, securityUserGetter.getAuthority().get(0));
+        return entityToOverviewDto(client);
+    }
+}
