+++ src/test/java/com/map/gaja/client/presentation/api/ClientPostControllerTest.java
@@ -1,10 +1,7 @@
 package com.map.gaja.client.presentation.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.map.gaja.client.application.ClientAccessVerifyService;
-import com.map.gaja.client.application.ClientBulkService;
-import com.map.gaja.client.application.ClientQueryService;
-import com.map.gaja.client.application.ClientService;
+import com.map.gaja.client.application.*;
 import com.map.gaja.client.domain.exception.InvalidFileException;
 import com.map.gaja.client.infrastructure.s3.S3FileService;
 import com.map.gaja.client.application.validator.ClientRequestValidator;
@@ -60,6 +57,8 @@ public class ClientPostControllerTest {
     ClientRequestValidator clientRequestValidator;
     @MockBean
     AuthenticationRepository authenticationRepository;
+    @MockBean
+    ClientSavingService clientSavingService;
 
     private ObjectMapper om;
 
@@ -97,10 +96,10 @@ public class ClientPostControllerTest {
         ClientOverviewResponse response = new ClientOverviewResponse();
         StoredFileDto fileDto = new StoredFileDto("test", "test");
         response.setImage(fileDto);
-        when(clientService.saveClientWithImage(any(),any())).thenReturn(response);
+        when(clientSavingService.saveClientWithImage(any(),any())).thenReturn(response);
 
         mvc.perform(mockRequest).andExpect(MockMvcResultMatchers.status().isPartialContent());
-        verify(clientService, times(1)).saveClientWithImage(any(), any());
+        verify(clientSavingService, times(1)).saveClientWithImage(any(), any());
     }
 
     @Test
@@ -115,10 +114,10 @@ public class ClientPostControllerTest {
         ClientOverviewResponse response = new ClientOverviewResponse();
         StoredFileDto fileDto = new StoredFileDto("test", "test");
         response.setImage(fileDto);
-        when(clientService.saveClientWithImage(any(),any())).thenReturn(response);
+        when(clientSavingService.saveClientWithImage(any(),any())).thenReturn(response);
 
         mvc.perform(mockRequest).andExpect(MockMvcResultMatchers.status().isCreated());
-        verify(clientService, times(1)).saveClientWithImage(any(), any());
+        verify(clientSavingService, times(1)).saveClientWithImage(any(), any());
     }
 
     @Test
@@ -129,10 +128,10 @@ public class ClientPostControllerTest {
         MockHttpServletRequestBuilder mockRequest = ClientRequestCreator.createPostRequestWithImage(testUrl);
         ClientRequestCreator.setNormalField(mockRequest, request);
         mockRequest.param("isBasicImage", String.valueOf(false));
-        when(clientService.saveClientWithImage(any(), any())).thenThrow(new GroupNotFoundException());
+        when(clientSavingService.saveClientWithImage(any(), any())).thenThrow(new GroupNotFoundException());
 
         mvc.perform(mockRequest).andExpect(MockMvcResultMatchers.status().isUnprocessableEntity());
-        verify(clientService, times(1)).saveClientWithImage(any(), any());
+        verify(clientSavingService, times(1)).saveClientWithImage(any(), any());
     }
 
     @Test
